requirements.txt
Django~=4.0.4
numpy~=1.22.3
vosk~=0.3.32
django-ckeditor~=6.4.0
PyDrive~=1.3.1
mammoth~=1.4.19
htmldocx~=0.0.6
